+++ odoo/js/loaders/backend.js
@@ -1,4 +1,4 @@
-// Copyright 2018-2020 Alexandre Díaz <dev@redneboa.es>
+// Copyright  Alexandre Díaz <dev@redneboa.es>
 // License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 odoo.define("terminal.loaders.Backend", function (require) {