+++ settings/options.js
@@ -1,5 +1,6 @@
 /* global browser, chrome */
-// Copyright 2019-2020 Alexandre Díaz
+// Copyright  Alexandre Díaz <dev@redneboa.es>
+// License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 (function () {
     "use strict";