+++ tests/__init__.py
@@ -1,0 +1,2 @@
+# Copyright  Alexandre Díaz <dev@redneboa.es>
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).