+++ tests/test_extension.py
@@ -1,5 +1,5 @@
-# Copyright 2020 Alexandre Díaz
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
+# Copyright  Alexandre Díaz <dev@redneboa.es>
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 import pytest
 import time
 import pathlib