+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+**8.5.1**
+
+```
+IMP: Command 'search': better output
+
+FIX: ParameterReader: quotes (issue #65)
+```
+
 **8.5.0**
 
 ```